% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lap_theme.R
\docType{data}
\name{lap_theme}
\alias{lap_theme}
\title{thematic leprechaun theme}
\format{
An object of class \code{bs_theme_with_preset} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
lap_theme
}
\value{
bslib theme
}
\description{
thematic leprechaun theme
}
\keyword{datasets}
