% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatter_display.R
\name{scatter_display_server}
\alias{scatter_display_server}
\title{scatter_display Server}
\usage{
scatter_display_server(id, var_inputs, aes_inputs)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{var_inputs}{variable inputs}

\item{aes_inputs}{aesthetic inputs}
}
\description{
scatter_display Server
}
\keyword{internal}
