% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatter_display.R
\name{scatter_displayUI}
\alias{scatter_displayUI}
\title{scatter_display UI}
\usage{
scatter_displayUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
scatter_display UI
}
\keyword{internal}
