% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{The scatter plot utility function}
\usage{
scatter_plot(df, x_var, y_var, col_var, alpha_var, size_var)
}
\arguments{
\item{df}{\code{data.frame} or \code{tibble}}

\item{x_var}{string variable mapped to \code{x} axis}

\item{y_var}{string variable mapped to \code{y} axis}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A custom graphing ggplot2 function
}
